requirements.txt
firebase-admin>=7.0.0
torch>=2.9.0
torchvision>=0.18.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
faiss-cpu>=1.7.4
mitreattack-python>=3.0.5
python-dotenv>=1.0.0
